requirements.txt
Flask==2.0.2
twilio==7.0.0
openai==0.27.0
requests==2.26.0
speech-recognition==3.8.1
whisper-autograd==0.2.1
